Allow several _ext -> header copy blocks, no longer automatically add all includes from _ext file to header, detect includes in copy to header blocks, rename util.hpp to warning_macros.hpp
