+++ rerun_cpp/src/rerun/archetypes/depth_image.hpp
@@ -8,10 +8,9 @@
 #include "../components/draw_order.hpp"
 #include "../components/tensor_data.hpp"
 #include "../data_cell.hpp"
-#include "../error.hpp"
 #include "../indicator_component.hpp"
 #include "../result.hpp"
-#include "../util.hpp"
+#include "../warning_macros.hpp"
 
 #include <cstdint>
 #include <optional>
