+++ rerun_cpp/src/rerun/archetypes/line_strips2d.hpp
@@ -14,7 +14,7 @@
 #include "../data_cell.hpp"
 #include "../indicator_component.hpp"
 #include "../result.hpp"
-#include "../util.hpp"
+#include "../warning_macros.hpp"
 
 #include <cstdint>
 #include <optional>
