+++ rerun_cpp/src/rerun/archetypes/line_strips3d.hpp
@@ -13,7 +13,7 @@
 #include "../data_cell.hpp"
 #include "../indicator_component.hpp"
 #include "../result.hpp"
-#include "../util.hpp"
+#include "../warning_macros.hpp"
 
 #include <cstdint>
 #include <optional>
