+++ rerun_cpp/src/rerun/archetypes/time_series_scalar.hpp
@@ -12,7 +12,7 @@
 #include "../data_cell.hpp"
 #include "../indicator_component.hpp"
 #include "../result.hpp"
-#include "../util.hpp"
+#include "../warning_macros.hpp"
 
 #include <cstdint>
 #include <optional>
