+++ rerun_cpp/src/rerun/components/line_strip2d.hpp
@@ -7,7 +7,6 @@
 #include "../datatypes/vec2d.hpp"
 #include "../result.hpp"
 
-#include <algorithm>
 #include <cstdint>
 #include <memory>
 #include <utility>
