+++ rerun_cpp/src/rerun/datatypes/annotation_info.hpp
@@ -10,7 +10,6 @@
 #include <cstdint>
 #include <memory>
 #include <optional>
-#include <utility>
 
 namespace arrow {
     class DataType;
