+++ rerun_cpp/src/rerun/datatypes/class_description_map_elem.hpp
@@ -9,7 +9,6 @@
 
 #include <cstdint>
 #include <memory>
-#include <utility>
 
 namespace arrow {
     class DataType;
