+++ rerun_cpp/src/rerun/datatypes/keypoint_pair.hpp
@@ -8,7 +8,6 @@
 
 #include <cstdint>
 #include <memory>
-#include <utility>
 
 namespace arrow {
     class DataType;
