+++ rerun_cpp/src/rerun/datatypes/tensor_buffer.hpp
@@ -6,7 +6,6 @@
 #include "../half.hpp"
 #include "../result.hpp"
 
-#include <cassert>
 #include <cstdint>
 #include <cstring>
 #include <memory>
