+++ rerun_cpp/src/rerun/datatypes/translation_rotation_scale3d_ext.cpp
@@ -1,20 +1,13 @@
-#include "../util.hpp"
 #include "translation_rotation_scale3d.hpp"
 
-// Uncomment for better auto-complete while editing the extension.
-// #define EDIT_EXTENSION
+// [CODEGEN COPY TO HEADER START]
+#include "../warning_macros.hpp"
 
+// [CODEGEN COPY TO HEADER END]
 namespace rerun {
     namespace datatypes {
 
-#ifdef EDIT_EXTENSION
-        struct TranslationRotationScale3DExt {
-            std::optional<Vec3D> translation;
-            std::optional<Rotation3D> rotation;
-            std::optional<Scale3D> scale;
-            bool from_parent;
-
-#define TranslationRotationScale3D TranslationRotationScale3DExt
+#if 0
             // [CODEGEN COPY TO HEADER START]
 
             static const TranslationRotationScale3D IDENTITY;
@@ -170,13 +163,9 @@ namespace rerun {
 
             // [CODEGEN COPY TO HEADER END]
         };
-
-#undef TranslationAndMat3x3
-#else
-#define TranslationRotationScale3DExt TranslationRotationScale3D
 #endif
 
-        const TranslationRotationScale3DExt TranslationRotationScale3DExt::IDENTITY =
+        const TranslationRotationScale3D TranslationRotationScale3D::IDENTITY =
             TranslationRotationScale3D();
 
     } // namespace datatypes
