+++ rerun_cpp/src/rerun/warning_macros.hpp
